(TEST)(FIX) all CUDA nn module and init tests pass

+++ hypercoil/nn/test/test_freqfilter.py
@@ -80,6 +80,6 @@ class TestFreqFilter:
     def test_cuda_forward(self):
         filt = FrequencyDomainFilter(
             self.filter_specs, dim=50,
-            device='cuda', dtype=torch.half)
-        out = filt(self.Z4.clone().cuda().half())
+            device='cuda', dtype=torch.double)
+        out = filt(self.Z4.clone().cuda().double())
         assert out.size() == torch.Size([1, 13, 7, 99])
